deleted all authentication and changed home some ui
