+++ front/src/App.jsx
@@ -1,96 +1,32 @@
 import './App.css';
-import { useState, useEffect } from 'react';
-import Login from './components/Users/Login';
-import SignUp from './components/Users/SignUp';
 import Home from './pages/home';
 import Accounts from './pages/accounts';
 import Transactions from './pages/transactions';
 import NewTransaction from './components/Transactions/NewTransaction/NewTransaction';
 import EditTransaction from './components/Transactions/EditTransaction';
-import { getUsers, addUser } from './lib/userData';
-import { BrowserRouter as Router, Route, Link, Routes, Navigate } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Link, Routes } from 'react-router-dom';
 
 function App() {
-	const [users, setUsers] = useState([]);
-	const [currentUser, setCurrentUser] = useState();
-	const [errorMessage, setErrorMessage] = useState('');
-	const [hideSignup, setHideSignup] = useState(false);
-
-	useEffect(() => {
-		async function getData() {
-			const userData = await getUsers();
-			setUsers(userData);
-		}
-		getData();
-	}, []);
-
-	function loginHandler(name, password) {
-		users.forEach((user) => {
-			if (user.name === name && user.password === password) {
-				setCurrentUser({ name, password });
-			} else {
-				setErrorMessage('Your name or password is incorrect!');
-			}
-		});
-	}
-
-	async function signUpHandler(name, password) {
-		await addUser(name, password);
-		const updatedUsers = await getUsers();
-
-		setUsers(updatedUsers);
-		setHideSignup(true);
-	}
-
 	return (
 		<Router>
 			<div className='App'>
-				<nav>
-					<ul>
-						{currentUser ? (
-							<div style={{ marginBottom: '10px' }}>
-								{' '}
-								Hello, {currentUser.name} <a href='#.com'>logout</a>
-							</div>
-						) : (
-							<>
-								<li>
-									<Link to='/login'>Login</Link>
-								</li>
-								{hideSignup ? null : (
-									<li>
-										<Link to='/signup'>Signup</Link>
-									</li>
-								)}
-							</>
-						)}
-						<Link to='/'>Home</Link>
-						<br />
-						<Link to='/transactions'>Transactions</Link>
+				<nav style={{ marginBottom: '50px' }}>
+					<ul style={{ display: 'flex' }}>
+						<Link style={{ marginRight: '10px' }} to='/'>
+							Home
+						</Link>
+						<Link style={{ marginRight: '10px' }} to='/transactions'>
+							Transactions
+						</Link>
+						<Link style={{ marginRight: '10px' }} to='/accounts'>
+							Accounts
+						</Link>
+						<Link to='/transactions/new'>Add new transaction</Link>
 					</ul>
 				</nav>
 
 				<Routes>
-					<Route
-						path='/login'
-						element={
-							currentUser ? (
-								<Navigate to='/' replace={true} />
-							) : (
-								<Login
-									onLogin={loginHandler}
-									errorMessage={errorMessage}
-									currentUser={currentUser}
-								/>
-							)
-						}
-					/>
-					<Route path='/signup' element={hideSignup ? null : <SignUp onSignUp={signUpHandler} />} />
-
-					<Route
-						path='/'
-						element={!currentUser ? <Navigate to='/login' replace={true} /> : <Home />}
-					/>
+					<Route path='/' element={<Home />} />
 					<Route path='/accounts' element={<Accounts />} />
 					<Route path='/transactions' element={<Transactions />} />
 					<Route path='/transactions/new' element={<NewTransaction />} />
