+++ front/src/components/Transactions/NewTransaction/TransactionForm.module.css
@@ -1,3 +1,4 @@
-.transfer div{
-    margin-bottom: 10px;
+
+.transfer {
+    margin-top: 10px
 }
