+++ front/src/components/Users/SignUp.jsx
@@ -1,48 +1,26 @@
 import { useState } from 'react';
+import Form from './Form';
 
 export default function SignUp({ onSignUp }) {
-    const [newName, setNewName] = useState('');
-    const [newPass, setNewPass] = useState('');
+	const [user, setUser] = useState({ name: '', password: '' });
 
-    function nameChangeHandler(event) {
-        setNewName(event.target.value);
-    }
+	function changeHandler(event) {
+		setUser((prevUser) => {
+			return {
+				...prevUser,
+				[event.target.name]: event.target.value,
+			};
+		});
+	}
 
-    function passChangeHandler(event) {
-        setNewPass(event.target.value);
-    }
+	function submitHandler(event) {
+		event.preventDefault();
+		onSignUp(user);
+	}
 
-    function submitHandler(e) {
-        e.preventDefault();
-        onSignUp(newName, newPass);
-    }
-
-    return (
-        <>
-            <h4>Create Account</h4>
-            <form onSubmit={submitHandler}>
-                <div style={{ marginBottom: '5px' }}>
-                    <label htmlFor="name">username: </label>
-                    <input
-                        value={newName}
-                        onChange={nameChangeHandler}
-                        type="text"
-                        id="name"
-                        name="username"
-                    />
-                </div>
-                <div style={{ marginBottom: '10px' }}>
-                    <label htmlFor="pass">password: </label>
-                    <input
-                        value={newPass}
-                        onChange={passChangeHandler}
-                        type="password"
-                        id="pass"
-                        name="password"
-                    />
-                </div>
-                <button>create account</button>
-            </form>
-        </>
-    );
+	return (
+		<>
+			<Form onFormSubmit={submitHandler} onInputChange={changeHandler} />
+		</>
+	);
 }
